Split test file up into two files
AI-assisted
