+++ .pylintrc
@@ -1,2 +1,5 @@
+[MASTER]
+init-hook='import sys; sys.path.append(".")'
+
 [MESSAGES CONTROL]
 disable=duplicate-code,R0801
